fix(cf): pinging same contest multiple times
Related issue: #36
Problem:
When iterating over upcomingContests a copy of the contest is created,
which makes it impossible to modify the contest stored inside
upcomingContests when iterating over it.

Solution:
Store upcomingContests as pointers to contests, when a copy if this
value is made it will still point to the same variable.