refactor: open session before adding message handler
This allows all initialization of managers to happen inside the
NewManager() function, because some of the initialization requires the
session to already be open.
Considering that managers must be created before adding the message
handler (otherwise the handler has no manager to call functions from),
the session is also opened before this.