feat(cf): track pinged contests by ID
Tracking pinged contests by ID fixes a possible issue where codeforces
contests are refetched, resetting the Pinged variable associated with
the contest, thus triggering a new ping. The ID of the refetched
contests will be the same as the previous ones, eliminating this issue.
Also insert new contests in a way that maintains the sorted order, for
more efficient contest ping checking.