feat(ci): add compose profiles and db healthcheck
Closes #55
Improves deployment by having different profiles inside a single compose
file, as well as making networking with the db easier, as everything is
on the same docker network regardless.
Also add healthcheck to the db service and bot container is only started
after this succeeds, to prevent unnecessary restarts of the container.