refactor: create Codeforces Service struct
The Codeforces package now has a Service struct responsible for all
state inside the package. In addition to improved readability, this also
has the benefit of technically allow for multiple of these to exist and
run.

Also created a new api.go file which is responsible for all code
directly interacting with the Codeforces API.