refactor: create interface for database
The codeforces package defines a Repository interface with all the
functions it needs, the db struct in the database package implements all
of these functions. This allows the codeforces package to only be able
to call their relevent functions, and allows mocking of the database.

Also change database functions to take a context parameter instead of
always using context.Background().