fix: auth timeout and bad connection check
Problem:
The timeout was not actually set inside the auth service, resulting in
it being 0. It was also a time.Time, when it should have been a
time.Duratin.
I had also forgotten a "!" in front of an if when checking if the user
is already connected.

Solution:
Set timeout in newAuthService as a Duration, and add "!" in front of if.